Fix no caps bug and use multiple dictionaries
Previously script was throwing away words that were used more often
capitalized than not. This turned out to be a mistake as it got rid of
lots of good words like "Abbey" and "University".

Also, script would regenerate the WordNet list, but did not check the
list against any other dictionaries. Now it tries wordnet, gcide, and
(if you have it installed) OED.