Update script for 2020 data format
I thought I had already done this, but I guess I never checked in the
change. The new format uses *COMMAS* as well as TABs as field
separators.

word	year,count,books	year,count,books	year,count,books
apple	1883,5,5	1884,48,27	1885,650,201	...

This is a bad idea because

(a): comma is a valid character in a word, so one can no longer simply
split on the field separators,

(b): there was no pressing need to demarcate these fields separately,
and

(c) there was a perfectly good alternative: SPACE.

Having to parse the fields in a for loop makes the preaccumulation
step take much longer.