Tool to replace commas with spaces in input data after first field
Google changed the data format to use commas as a subfield separator.
That was silly since comma is a valid character in data in the first
field, which means one cannot create regexp to easily split the line
into fields.

This fixes the format by replacing commas with spaces *iff* the comma
comes after a TAB on the line.