Add number of books by year and sort by that
Previously we threw away the book count and only used the total number
of occurrences. Unfortunately, this made certain words seem more
common ("psychosocial") because they were used incessantly in a few
specialized books.

This is the first step to making a metric which combines both
occurrences and books to determine "popularity".